+++ extensions-contrib/range/src/test/java/io/druid/query/aggregation/range/MetricRangeTest.java
@@ -15,17 +15,17 @@ public class MetricRangeTest
   {
     MetricRange mr = new MetricRange();
 
-    float[] data = {0.1f, 0.8f, 0.3f, 0.6f, 0.5f, 0.2f, 0.1f, 0.7f};
+    double[] data = {0.1, 0.8, 0.3, 0.6, 0.5, 0.2, 0.1, 0.7};
 
-    for (float val : data)
+    for (double val : data)
     {
       mr.add(val);
     }
 
-    MetricRange expected = new MetricRange(0.1f, 0.8f);
+    MetricRange expected = new MetricRange(0.1, 0.8);
 
     Assert.assertEquals(expected, mr);
-    Assert.assertEquals(0.7f, mr.getRange(), 0.001);
+    Assert.assertEquals(0.7, mr.getRange(), 0.001);
   }
 
   @Test
